refactor: Move data files into a /app/data dir
