+++ app/layout.tsx
@@ -3,7 +3,7 @@ import "./globals.css";
 import React, { useEffect, useState } from "react";
 import { Analytics } from "@vercel/analytics/react";
 import { Inter } from "next/font/google";
-import { organizationsStore } from "./organizations";
+import { organizationsStore } from "./data/organizations";
 import {
   NextUIProvider,
   Navbar,
